Feature/montage (#3)
* Added montage and tile index classes for both Grid montages and the abstract base class for creating other types of montages.
* Fixed Mac and Linux compile issues.
* Added additional unit tests for the GridMontage class.

+++ src/complex/DataStructure/DataObject.hpp
@@ -35,7 +35,7 @@ public:
    * @brief The IdType alias serves as an ID type for DataObjects within their
    * respective DataStructure.
    */
-  using IdType = size_t;
+  using IdType = uint64_t;
 
   /**
    * @brief The ParentCollectionType alias describes the format by which parent
